requirements.txt
Django==1.11.3
django-mptt==0.8.7
mysqlclient==1.3.10
olefile==0.44
Pillow==4.2.1
pkg-resources==0.0.0
python-slugify==1.2.4
pytz==2017.2
Unidecode==0.4.21
